moved template caching to compile step
Saving the template during the linking step causes the problem that some directives might already have modified the DOM (e.g., ng-repeat) and in such a case a wrong template is saved. For the case of ng-repeat, it causes that the repeat is not executed anymore. Therefore, I moved this task to the compile step where the data is still clean.